+++ RowDaBoat-x64barebones/Kernel/interrupts/int80.h
@@ -42,7 +42,7 @@ void* MyMalloc(uint64_t rsi);// sys call 14
 void MyFree(void* rsi);// sys call 15
 void* memSet(void* rsi,uint32_t rdx,uint64_t rcx);//sys call 16
 void mem(mm_stat*rsi);// sys call 17
-int exec(main_func_t * rsi, char* rdx, int rcx, int * r8, int* r9); //18
+int exec(main_func_t * rsi, char* rdx, int rcx, int * r8, uint64_t* r9); //18
 void pYield();
 int ps(process_info * rsi, unsigned int rdx, unsigned int * rcx);// sys call 19
 int process_kill(int rsi);// sys call 21
